requirements.txt
Django==5.1.6
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
mysqlclient==2.1.1
psycopg2-binary==2.9.6
python-dotenv==1.0.0
dnspython==2.7.0